refactor(database): extract byte-identical wrappers + test loggers
Two byte-identical code clusters that lived in both
database/postgresql/ and database/oracle/ are now centralized in
database/internal/ subpackages. Both vendor packages re-export the
Statement and Transaction types via type aliases, preserving the
public API surface.

## Extracted to database/internal/wrapper/

- Statement (wraps sql.Stmt, implements types.Statement)
- Transaction (wraps sql.Tx, implements types.Tx)

These were 67 LOC each in postgresql/connection.go and
oracle/connection.go — pre-refactor, byte-identical except for one
"PostgreSQL" vs "Oracle" comment. Now ~85 LOC in one place. The
struct fields (stmt, tx) become unexported in the wrapper package, so
callers must use NewStatement / NewTransaction constructors instead
of struct literals — already updated at all 7 internal callsites
(Connection.Prepare, Connection.Begin, Connection.BeginTx in both
packages, plus the within-Transaction Prepare path).

## Extracted to database/internal/dbtestlog/

- NewTestLogger (was newTestLogger in both packages)
- NewDisabledTestLogger (was newDisabledTestLogger in both packages)

These were 18 LOC each in testhelpers.go files — byte-identical
except for the package declaration and one "PostgreSQL" vs "Oracle"
godoc word. Now lives in database/internal/dbtestlog/log.go,
exported (capital N) so both packages can import. The old
testhelpers.go files are deleted.

## Net effect

- 2 new internal packages (wrapper, dbtestlog) — ~110 LOC each
- 2 vendor files trimmed by ~67 LOC each (Statement/Transaction body)
- 2 testhelpers.go files deleted (-18 LOC each)
- Test files migrated to use the new constructors and helpers

Diff stat: -230 lines / +82 lines = net -148 LOC of duplication
removed.

## Why type aliases (and not interface or move-and-rename)

postgresql.Statement and oracle.Statement are part of the established
public API. A move-and-rename would break framework consumers; an
interface would force a runtime-dispatch boundary that has no value.
Type aliases (type X = pkg.Y) preserve full backwards compatibility
— a caller using postgresql.Statement still gets exactly the same
type as wrapper.Statement, with zero indirection.

Wave 4-B — first of the database dedup pair (W4-C will extract the
connection delegation methods next).

Pre-push: make check (43 packages clean, race), gosec clean,
govulncheck clean, integration build clean.
